Lock docker

+++ central_server/src/main/java/com/example/centralServer/FileUploadController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -14,6 +15,8 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
+//lock
+import com.example.centralServer.redisLock.RedisLockTool;
 
 import java.io.IOException;
 
@@ -22,6 +25,8 @@ import static org.springframework.web.bind.annotation.RequestMethod.POST;
 @Controller
 public class FileUploadController {
     ServerManager serverManager;
+    @Autowired
+    RedisLockTool lockManager;
 
     @Autowired
     public FileUploadController(ServerManager serverManager) {
@@ -36,10 +41,11 @@ public class FileUploadController {
      */
     @GetMapping("/dirs")
     @ResponseBody
-    public ResponseEntity<?> getDirectoriesByPath(@RequestParam(value = "path", defaultValue = "/") String path) {
-        String url = serverManager.getNextAvailableServerUrl();
-        RestTemplate restTemplate = new RestTemplate();
-        return restTemplate.getForEntity(url + "/dirs?path=" + path, String.class);
+    public ResponseEntity<?> getDirectoriesByPath(@RequestParam(value = "path", defaultValue = "/") String path) { 	
+    	String url = serverManager.getNextAvailableServerUrl();
+        RestTemplate restTemplate = new RestTemplate();     
+        ResponseEntity response= restTemplate.getForEntity(url + "/dirs?path=" + path, String.class);
+        return response;
 
     }
 
@@ -52,9 +58,22 @@ public class FileUploadController {
     @PostMapping("/dirs")
     @ResponseBody
     public ResponseEntity<?> createDirectoryByPath(@RequestParam(value = "path") String path) {
-        String url = serverManager.getNextAvailableServerUrl();
+    	//lock
+    	String lockKey = path;
+    	String identity = "deleteDirLock";
+    	boolean lockSuccess = lockManager.getLock(lockKey, identity, 60);
+    	if (lockSuccess) {
+    	String url = serverManager.getNextAvailableServerUrl();
         RestTemplate restTemplate = new RestTemplate();
-        return restTemplate.postForEntity(url + "/dirs?path=" + path, null, String.class);
+        ResponseEntity response = restTemplate.postForEntity(url + "/dirs?path=" + path, null, String.class);
+    
+      //release lock
+        boolean releaseSuccess = (boolean) lockManager.unlock(lockKey, identity);\
+        
+        return response;
+    	}else {
+     		return ResponseEntity.status(HttpStatus.CONFLICT).body("Directory is bing modified");
+    	}
     }
 
 
@@ -67,9 +86,18 @@ public class FileUploadController {
     @DeleteMapping("/dirs")
     @ResponseBody
     public void deleteDirectoryByPath(@RequestParam(value = "path") String path) {
-        String url = serverManager.getNextAvailableServerUrl();
+    	//lock
+    	String lockKey = path;
+    	String identity = "deleteDirLock";
+    	boolean lockSuccess = lockManager.getLock(lockKey, identity, 60);
+    	
+    	if (lockSuccess) {
+    	String url = serverManager.getNextAvailableServerUrl();
         RestTemplate restTemplate = new RestTemplate();
         restTemplate.delete(url + "/dirs?path=" + path);
+      //release lock
+        boolean releaseSuccess = (boolean) lockManager.unlock(lockKey, identity);
+    	}
     }
 
     /**
@@ -80,11 +108,12 @@ public class FileUploadController {
      */
     @GetMapping("/files")
     @ResponseBody
-    public ResponseEntity<Resource> getFileByPath(@RequestParam(value = "path") String path) {
-        String url = serverManager.getNextAvailableServerUrl();
+    public ResponseEntity<?> getFileByPath(@RequestParam(value = "path") String path) {
+    	String url = serverManager.getNextAvailableServerUrl();
         RestTemplate restTemplate = new RestTemplate();
         return restTemplate.getForEntity(url + "/files?path=" + path, Resource.class);
     }
+    
 
     /**
      * upload a file in a multipart form
@@ -94,7 +123,13 @@ public class FileUploadController {
      */
     @RequestMapping(path = "/files", method = POST, consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
     public ResponseEntity<?> handleFileUpload(@RequestBody MultipartFile file, @RequestParam("path") String path) throws IOException {
-        MultiValueMap<String, Object> map = new LinkedMultiValueMap<>();
+    	  //lock
+    	String lockKey = path;
+    	String identity = "uploadFileLock";
+    	boolean lockSuccess = lockManager.getLock(lockKey, identity, 60);
+    	if (lockSuccess) {
+    	
+    	MultiValueMap<String, Object> map = new LinkedMultiValueMap<>();
         map.add("file", file.getResource());
 
         HttpHeaders headers = new HttpHeaders();
@@ -106,7 +141,13 @@ public class FileUploadController {
         RestTemplate restTemplate = new RestTemplate();
 
         String result = restTemplate.postForObject(url + "/files?path=" + path, requestEntity, String.class);
+      //release lock
+        boolean releaseSuccess = (boolean) lockManager.unlock(lockKey, identity);
+        
         return ResponseEntity.ok(result);
+    	}else {
+     		return ResponseEntity.status(HttpStatus.CONFLICT).body("Directory is bing modified");
+    	}
     }
 
 
@@ -118,9 +159,17 @@ public class FileUploadController {
     @DeleteMapping("/files")
     @ResponseBody
     public void deleteFileByPath(@RequestParam(value = "path") String path) {
+    	 //lock
+    	String lockKey = path;
+    	String identity = "deltetFileLock";
+    	boolean lockSuccess = lockManager.getLock(lockKey, identity, 60);
+    	if (lockSuccess) {
         String url = serverManager.getNextAvailableServerUrl();
         RestTemplate restTemplate = new RestTemplate();
         restTemplate.delete(url + "/files?path=" + path);
+        //release lock
+        boolean releaseSuccess = (boolean) lockManager.unlock(lockKey, identity);
+    	}
     }
 
 }
